move ./src/plugins/layers/satelliteOrbit.js to ./src/utils/orbit.js
